+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 New features and improvements:
 * Added `snap` command (#110)
 * Invisible SVG elements are now discarded (#103)
+* Add support for angle units (affects `rotate`, `skew`, and `arc` commands, `--radian` option is removed) (#111)
 
 API changes:
-* Added vpype_cli.execute() to execute a vpype pipeline from python (#104)
+* Added `vpype_cli.execute()` to execute a vpype pipeline from Python (#104)
+* Added `vpype.convert_angle()` and `vpype.AngleType` (#111)
 
 Bug fixes:
 * Fixed `write` to cap SVG width and height to a minimum of 1px (#102)